pyproject.toml
[tool.poetry]
name = "pplx-cli"
version = "0.1.0"
description = "A command-line interface for Perplexity AI API"
authors = ["xerexcoded <arnavbanerjee61@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/xerexcoded/pplx-cli"
repository = "https://github.com/xerexcoded/pplx-cli"
documentation = "https://github.com/xerexcoded/pplx-cli#readme"
keywords = ["perplexity", "ai", "cli", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.9.0"
requests = "^2.31.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
perplexity = "pplx_cli.cli:app"
